Update setup script.

+++ setup_watchdog.sh
@@ -4,9 +4,10 @@ set -e
 set -o pipefail
 set -o nounset
 
-sudo apt-get update
-
-sudo apt-get install -y watchdog
+if ! dpkg -s watchdog >/dev/null; then
+  sudo apt-get update
+  sudo apt-get install -y watchdog
+fi
 
 if ! cat /etc/modules | grep -q bcm2835_wdt; then
   echo 'bcm2835_wdt' | sudo tee -a /etc/modules
@@ -22,6 +23,8 @@ gateway_ip="$(ip route | awk '/^default via ([^\s]+) / { print $3 }')"
 
 cat <<CONFIG | sudo tee /etc/watchdog.conf
 watchdog-device	= /dev/watchdog
+watchdog-timeout = 10
+interval = 2
 max-load-1 = 24
 ping = ${gateway_ip}
 CONFIG
